+++ feincms3/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (0, 17, 3)
+VERSION = (0, 18, 0)
 __version__ = '.'.join(map(str, VERSION))